+++ FormSystem/FormSystem/FormSystem.csproj
@@ -128,6 +128,7 @@
     <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
+    <Compile Include="Controllers\BackStageController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="DBModel\FormData.cs" />
     <Compile Include="DBModel\FormDBModel.cs" />
@@ -147,6 +148,7 @@
     <Compile Include="Models\CreateFormModel.cs" />
     <Compile Include="Models\FrequenModel.cs" />
     <Compile Include="Models\IndexPagerModel.cs" />
+    <Compile Include="Models\LoginModel.cs" />
     <Compile Include="Models\ShowAnsModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -233,8 +235,8 @@
     <Content Include="Views\Home\Index.cshtml" />
     <Content Include="Views\Home\Login.cshtml" />
     <Content Include="Views\Home\FillForm.cshtml" />
-    <Content Include="Views\Home\FormManager.cshtml" />
-    <Content Include="Views\Home\FrequentlyQuestions.cshtml" />
+    <Content Include="Views\BackStage\FormManager.cshtml" />
+    <Content Include="Views\BackStage\FrequentlyQuestions.cshtml" />
     <Content Include="Views\Home\CheckAns.cshtml" />
     <Content Include="Views\PartialView\_NewInfo.cshtml" />
     <Content Include="Views\PartialView\_NewLayout.cshtml" />
