+++ FormSystem/FormSystem/Views/Shared/_Layout.cshtml
@@ -21,8 +21,8 @@
                 <div class="navbar-nav">
                     <a class="nav-link">@Html.ActionLink("首頁", "Index", "Home")</a>
                     <a class="nav-link">@Html.ActionLink("管理員登入", "Login", "Home")</a>
-                    <a class="nav-link">@Html.ActionLink("問卷管理", "FormManager", "Home")</a>
-                    <a class="nav-link">@Html.ActionLink("常用問題管理", "FrequentlyQuestions", "Home")</a>
+                    <a class="nav-link">@Html.ActionLink("問卷管理", "FormManager", "BackStage")</a>
+                    <a class="nav-link">@Html.ActionLink("常用問題管理", "FrequentlyQuestions", "BackStage")</a>
                 </div>
             </div>
         </div>
