+++ FormSystem/FormSystem/Web.config
@@ -17,6 +17,9 @@
   <system.web>
     <compilation debug="true" targetFramework="4.8" />
     <httpRuntime targetFramework="4.8" />
+	<authentication mode="Forms">
+		<forms loginUrl="/Home/Login" timeout="2880" />
+	</authentication>
   </system.web>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
